requirements.txt
python-telegram-bot==13.1
pandas
requests
PyYAML
